Split components into standalone files
