+++ examples/simple/src/comments/index.js
@@ -1,0 +1,13 @@
+import ChatBubbleIcon from '@material-ui/icons/ChatBubble';
+import CommentCreate from './CommentCreate';
+import CommentEdit from './CommentEdit';
+import CommentList from './CommentList';
+import CommentShow from './CommentShow';
+
+export default {
+    list: CommentList,
+    create: CommentCreate,
+    edit: CommentEdit,
+    show: CommentShow,
+    icon: ChatBubbleIcon,
+};
