+++ examples/simple/src/posts/ResetViewsAction.js
